+++ fabric/src/main/java/net/SpectrumFATM/black_archive/fabric/entity/client/ModModelLayers.java
@@ -8,4 +8,5 @@ public class ModModelLayers {
     public static final EntityModelLayer DALEK_SLAVE = new EntityModelLayer(new Identifier("black_archive", "dalek_slave"), "main");
     public static final EntityModelLayer CYBERMAN = new EntityModelLayer(new Identifier("black_archive", "cyberman"), "main");
     public static final EntityModelLayer CYBERMAT = new EntityModelLayer(new Identifier("black_archive", "cybermat"), "main");
+    public static final EntityModelLayer TIME_FISSURE = new EntityModelLayer(new Identifier("black_archive", "time_fissure"), "main");
 }